requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pytest==7.4.2
numpy==1.24.3
click==8.1.7